requirements.txt
# Core
tensorflow>=2.10.0
numpy>=1.23.0
scipy>=1.9.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
pandas>=1.5.0
tqdm>=4.64.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.15.0

# Optional: GPU support
# tensorflow-gpu>=2.10.0  # Use if not using tensorflow with CUDA built-in
